refactor(cef): extract macos_compat.rs and logging.rs from lib.rs
Tier 3 item from docs/reports/REPORT_LARGE_FILE_MODULARIZATION_SCAN_2026_07_22.md.
Scoped narrowly per the audit's caution: pub fn run() itself (the CEF host
bootstrap, ~960+ lines with explicit ordering-hazard comments — macOS
sandbox init must precede framework load, etc.) was NOT touched or
restructured; only the two genuinely self-contained pieces it calls into
were relocated.

macos_compat.rs: the macOS ObjC runtime-patch cluster (patch_nsapp_
unrecognized_selector, disable_macos_drag_slideback, set_macos_app_
display_name variants, install_macos_accessibility_governor,
set_macos_activation_policy_regular, set_macos_dock_icon) — each
independently doc-commented, zero shared state with run() beyond being
called from it. logging.rs (bonus): init_logging + cleanup_old_logs.

lib.rs: 1949 -> 1121 lines. run()'s control flow, ordering, and comments
are unchanged beyond 8 call-site module-prefix additions. cargo build -p
agentmux-cef clean; cargo test -p agentmux-cef — 202 passed, 0 failed,
same count as before.

<!-- agentmux:agent_id=agentx -->
